src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/layout-tests-can-access-shadow-expected.txt

CONSOLE ERROR: line 28: Uncaught TypeError: p.createShadowRoot is not a function
This tests that layout tests can access shadow DOM.


PASS shadow.nodeName is "#document-fragment"
PASS internals.shadowRoot(shadow) threw exception TypeError: Failed to execute 'shadowRoot' on 'Internals': parameter 1 is not of type 'Element'..
PASS internals.shadowRoot(p) is null
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
